requirements.txt
Pygithub
python-multibar
disnake-jishaku
icecream
quart
PyYAML
requests
git+https://github.com/DisnakeDev/disnake
git+https://github.com/DisnakeDev/disnake#egg=disnake[discord,voice,speed]
firebase
asyncio
wavelink
CoffeePaginator
youtube_dl
psutil
opencv-python
pycryptodome
python_jwt
sseclient
gcloud
python-dotenv
requests_toolbelt
hypercorn
aiohttp
requests
python-socketio
Quart-Discord
discord-together
git+https://github.com/kuzaku-developers/discord-ext-dashboard.git
python-uptimerobot
Pillow
numpy
jishaku
websocket-client
python-socketio
termcolor
colorama
